Tests and refactoring
- Add some basic RSpec tests and run them
- Split modules into separate files
