+++ Gemfile
@@ -1,0 +1,6 @@
+source "https://rubygems.org"
+
+group :test do
+  gem "rspec"
+  gem "simplecov"
+end
